requirements.txt
torch>=2.0.0
numpy
h5py
pyyaml
onnx
onnxruntime
